Replaced FloatType by double and marked it as depricated.
git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@21211 b38c6c8f-bee8-465d-8448-9abe6915d496

+++ tracking/modules/trackFinderCDC/include/TrackFinderCDCAxialSegmentHoughModule.h
@@ -179,7 +179,7 @@ namespace Belle2 {
         // Check if the circle has been fitted reverse to the hough box by accident
         {
           GeneralizedCircle circle = trajectory2D.getGlobalCircle();
-          const FloatType& curv = circle.curvature();
+          const double& curv = circle.curvature();
           const std::array<DiscreteCurv, 2>& curvs = foundHoughBox.getBounds<DiscreteCurv>();
           const float& lowerCurv = *(curvs[0]);
           const float& upperCurv = *(curvs[1]);
